requirements.txt
telethon
pandas
dbt-core
dbt-postgres
dagster
dagster-postgres
ultralytics
fastapi
uvicorn
python-dotenv
psycopg2-binary
tqdm